+++ Source/HotspotMap.h
@@ -31,14 +31,14 @@ void setEnabledHotspotArrays(const BitVector<32>& theHotspotArrays);
 const BitVector<32>& getEnabledHotspotArrays();
 
 // Returns which hotspot to jump to in given direction (or 0)
-u16 getNextHotspotInDir(ECommandDir theDirection);
+int getNextHotspotInDir(ECommandDir theDirection);
 
 // Get a pre-generated map linking hotspots in a specific array
 // (with values relative to first hotspot in that array)
 struct ZERO_INIT(HotspotLinkNode)
 { u16 next[eCmdDir_Num]; bool edge[eCmdDir_Num]; };
 typedef std::vector<HotspotLinkNode> Links;
-const Links& getLinks(u16 theArrayID);
+const Links& getLinks(int theArrayID);
 
 // Converts a Profile string into a Hotspot (and removes the hotspot
 // from the start of the string in case there are multiple included).
